Added files via upload
__init__

well, this is a start to something I have always wanted to make.  So far the game is limited in things to do as there are only a few starting ships that roam aimlessly.  I am hoping to add multiplayer but that will take some time.  Good stuff.  Oh yeah, this was made using Python 3 with Pygame.